requirements.txt
fastapi==0.103.2
httpx==0.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.22
uvicorn==0.23.2